pyproject.toml
[project]
name = "python-seed"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.12.15",
    "azure-cosmos==4.9.0",
    "fastapi[standard-no-fastapi-cloud-cli]==0.116.1",
    "poethepoet==0.37.0",
    "pydantic==2.11.9",
    "pydantic-settings[azure-key-vault]==2.10.1",
]

[dependency-groups]
dev = [
    "jupyter==1.1.1",
    "pyright[nodejs]==1.1.405",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.0",
    "ruff==0.13.0",
]

[build-system]
requires = ["uv_build>=0.8.16,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
required-version = ">=0.8.16,<0.9.0"

[tool.pyright]
typeCheckingMode = "strict"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", # missing-trailing-comma
    "D100",   # undocumented-public-module
    "D101",   # undocumented-public-class
    "D102",   # undocumented-public-method
    "D103",   # undocumented-public-function
    "D104",   # undocumented-public-package
    "D105",   # undocumented-magic-method
    "D106",   # undocumented-public-nested-class
    "D107",   # undocumented-public-init
    "E501",   # line-too-long
    "ICN",    # flake8-import-conventions
    "S101",   # assert
    "TRY003", # raise-vanilla-args
]


[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["unit: Unit tests", "integration: Integration tests"]

[tool.poe.tasks]
check-code.shell = "ruff check && ruff format --diff && pyright"
start-api-dev = "fastapi dev src/python_seed/api/main.py"
start-api-pro.shell = "workers=$(nproc) && fastapi run --workers $workers src/python_seed/api/main.py"
